+++ src/settings.rs
@@ -13,12 +13,14 @@ const SETTINGS_FILENAME: &str = "settings.json";
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Settings {
     pub workrave_historystats_path: Option<String>,
+    pub workrave_todaystats_path: Option<String>,
 }
 
 impl Settings {
     pub fn default() -> Self {
         let mut settings = Self {
             workrave_historystats_path: None,
+            workrave_todaystats_path: None,
         };
         settings.init();
         settings
@@ -28,10 +30,11 @@ impl Settings {
         match fs::File::open(SETTINGS_FILENAME) {
             Ok(_) => {
                 self.load_settings().unwrap();
-            },
+            }
             Err(_) => {
                 println!("No settings file found");
-                self.workrave_historystats_path = Settings::try_workrave_appdata_path();
+                self.workrave_historystats_path = Settings::try_workrave_appdata_path("historystats");
+                self.workrave_todaystats_path = Settings::try_workrave_appdata_path("todaystats");
                 self.save_settings().unwrap();
             }
         };
@@ -44,8 +47,9 @@ impl Settings {
                 file.read_to_string(&mut contents)?;
                 let settings: Settings = serde_json::from_str(&contents)?;
                 self.workrave_historystats_path = settings.workrave_historystats_path;
+                self.workrave_todaystats_path = settings.workrave_todaystats_path;
                 Ok(true)
-            },
+            }
             Err(_) => {
                 println!("No existing settings file, saving current settings...");
                 match self.save_settings() {
@@ -55,15 +59,15 @@ impl Settings {
                     }
                 };
                 Ok(false)
-            },
+            }
         }
     }
 
-    fn try_workrave_appdata_path() -> Option<String> {
-        println!("Looking for historystats in default AppData path...");
-        let path = format!("C:\\Users\\{}\\AppData\\Roaming\\Workrave\\historystats", whoami::username());
+    fn try_workrave_appdata_path(filename: &str) -> Option<String> {
+        println!("Looking for {} in default AppData path...", filename);
+        let path = format!("C:\\Users\\{}\\AppData\\Roaming\\Workrave\\{}", whoami::username(), filename);
         if workrave::WorkraveHistory::is_file_valid(&path) {
-            println!("Found");
+            println!("Found {}", filename);
             Some(path)
         } else {
             None
@@ -79,12 +83,12 @@ impl Settings {
                     Ok(new_file) => {
                         println!("Settings file created");
                         new_file
-                    },
+                    }
                     Err(e) => panic!("Unknown failure when creating settings file: {:?}", e)
                 },
-                other_error=> {
+                other_error => {
                     panic!("Unknown failure when opening settings file: {:?}", other_error)
-                },
+                }
             }
         };
 
@@ -100,34 +104,63 @@ pub struct SettingsTab {
 }
 
 impl SettingsTab {
+    fn filepath_ui(ui: &mut Ui, path: &Option<String>) {
+        if let Some(saved_path) = &path {
+            ui.code(saved_path);
+        } else {
+            ui.code("! no path set !");
+        }
+    }
+
+    fn file_selection_dialog(ui: &mut Ui, dialog_title: &str, file_name: &str) -> Option<String> {
+        if ui.button("Set Path").clicked() {
+            Some(rfd::FileDialog::new()
+                .set_title(dialog_title)
+                .set_file_name(file_name)
+                .pick_file()?
+                .display()
+                .to_string())
+        } else {
+            None
+        }
+    }
+
     pub fn ui(&mut self, ui: &mut Ui) {
+        let mut have_settings_changed = false;
         ui.vertical(|ui| {
             ui.heading(RichText::new("Workrave"));
-            ui.horizontal(|ui| {
-                ui.label("\"historystats\" filepath: ");
-
-                if let Some(saved_path) = &self.settings.workrave_historystats_path {
-                    ui.monospace(saved_path);
-                } else {
-                    ui.monospace("! no path set !");
-                }
-
-                if ui.button("Set").clicked() {
-                    if let Some(path) = rfd::FileDialog::new()
-                        .set_title(format!("Select a Workrave \"{}\" file", workrave::WORKRAVE_HISTORYSTATS_FILENAME).as_str())
-                        .set_file_name(workrave::WORKRAVE_HISTORYSTATS_FILENAME.as_str())
-                        .pick_file() {
-                        let path = path.display().to_string();
+            ui.separator();
+            ui.columns(3, |columns| {
+                columns[0].vertical(|ui| {
+                    ui.label("historystats filepath");
+                    ui.separator();
+                    ui.label("todaystats filepath");
+                });
+                columns[1].vertical(|ui| {
+                    SettingsTab::filepath_ui(ui, &self.settings.workrave_historystats_path);
+                    ui.separator();
+                    SettingsTab::filepath_ui(ui, &self.settings.workrave_todaystats_path);
+                });
+                columns[2].vertical(|ui| {
+                    if let Some(path) = SettingsTab::file_selection_dialog(ui,
+                                                       format!("Select a Workrave \"{}\" file", workrave::WORKRAVE_HISTORYSTATS_FILENAME).as_str(),
+                                                       workrave::WORKRAVE_HISTORYSTATS_FILENAME) {
                         self.settings.workrave_historystats_path = Some(path);
+                        have_settings_changed = true;
                     }
-                }
+                    ui.separator();
+                    if let Some(path) = SettingsTab::file_selection_dialog(ui,
+                                                                           format!("Select a Workrave \"{}\" file", workrave::WORKRAVE_TODAYSTATS_FILENAME).as_str(),
+                                                                           workrave::WORKRAVE_TODAYSTATS_FILENAME) {
+                        self.settings.workrave_todaystats_path = Some(path);
+                        have_settings_changed = true;
+                    }
+                });
             });
-            if ui.button("Save settings").clicked() {
-                self.settings.save_settings().unwrap();
-            };
-            if ui.button("Load settings").clicked() {
-                self.settings.load_settings().unwrap();
-            }
         });
+
+        if have_settings_changed {
+            self.settings.save_settings().unwrap();
+        }
     }
 }